put the GlobalSnackbar in a separate file
